finish all routes

+++ src/routes.js
@@ -26,8 +26,16 @@ export const routes = [
     handler: (req, res) => {
       const { description, title } = req.body;
 
-      if (!description || !title) {
-        return res.writeHead(400).end();
+      if (!title) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: 'title not specified.' }));
+      }
+
+      if (!description) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: 'description not specified.' }));
       }
 
       const task = {
@@ -49,9 +57,61 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params;
-      const { name, email } = req.body;
+      const { description, title } = req.body;
 
-      database.update('tasks', id, { name, email });
+      if (!description && !title) {
+        return res
+          .writeHead(400)
+          .end(
+            JSON.stringify({ message: 'title or description are required.' })
+          );
+      }
+
+      const [task] = database.select('tasks', { id });
+
+      if (!task) {
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: 'task not found.' }));
+      }
+
+      const updatedTaskData = {
+        description,
+        title,
+        updated_at: new Date(),
+      };
+
+      if (!description) {
+        delete updatedTaskData.description;
+      }
+
+      if (!title) {
+        delete updatedTaskData.title;
+      }
+
+      database.update('tasks', id, updatedTaskData);
+
+      return res.writeHead(204).end();
+    },
+  },
+  {
+    method: 'PATCH',
+    path: buildRoutePath('/tasks/:id/complete'),
+    handler: (req, res) => {
+      const { id } = req.params;
+
+      const [data] = database.select('tasks', { id });
+
+      if (!data) {
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: 'task not found.' }));
+      }
+
+      const isTaskCompleted = data.completed_at !== null;
+      const completed_at = isTaskCompleted ? null : new Date();
+
+      database.update('tasks', id, { completed_at });
 
       return res.writeHead(204).end();
     },
@@ -61,6 +121,14 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params;
+
+      const [data] = database.select('tasks', { id });
+
+      if (!data) {
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: 'task not found.' }));
+      }
 
       database.delete('tasks', id);
